# Avoid creating new slices for labels values on postings for matchers