@machine424 2024-04-19
Thanks.
I suggest keeping `res` for enhanced readability and to avoid confusion:

```
	res := vals[:0]
	for _, val := range vals {
		if m.Matches(val) {
			res = append(res, val)
		}
	}

	if len(res) == 0 {
		return index.EmptyPostings(), nil
	}

	return ix.Postings(ctx, m.Name, res...)
```
This should maintain the same performance. WDYT?

Should we also adjust `inversePostingsForMatcher` as well?