@alanprot 2024-04-19
Great suggestion... I updated the PR!
The benchmark results remains the same!