@alanprot 2024-04-19
>Interesting, I'll dig into that when I have some time.

I think this is probably due the slice indexing.. im not sure... but is better anyway so all good i guess. :D