@alanprot 2024-04-19
We can get even better results if we pool the slices from `LabelValues` func [here](https://github.com/prometheus/prometheus/blob/f36915b6b14e3b54132a9cd701835d0255741f96/tsdb/index/postings.go#L149) and [here](https://github.com/prometheus/prometheus/blob/f36915b6b14e3b54132a9cd701835d0255741f96/tsdb/index/index.go#L1539)

WIP change: https://github.com/alanprot/prometheus/commit/e279a8352d3a2c9e557079d454a3b56c5c9a0f2f

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
                                             │  /tmp/old   │              /tmp/new              │
                                             │   sec/op    │   sec/op     vs base               │
Querier/Head/PostingsForMatchers/i=~"1.*"-32   8.131m ± 2%   6.747m ± 1%  -17.02% (p=0.002 n=6)

                                             │   /tmp/old   │              /tmp/new               │
                                             │     B/op     │     B/op      vs base               │
Querier/Head/PostingsForMatchers/i=~"1.*"-32   3.357Mi ± 0%   1.212Mi ± 2%  -63.90% (p=0.002 n=6)

                                             │  /tmp/old   │             /tmp/new              │
                                             │  allocs/op  │  allocs/op   vs base              │
Querier/Head/PostingsForMatchers/i=~"1.*"-32   11.14k ± 0%   11.12k ± 0%  -0.13% (p=0.002 n=6)
```

@yeya24  @machine424 Do you think is worth it? (i think it is as it shows a big improvement on memory allocation and probably it will have a impact on GC if we run multiples regex queries)

If so i can create a follow up PR.