@machine424 2024-04-23
> @yeya24 @machine424 Do you think is worth it? (i think it is as it shows a big improvement on memory allocation and probably it will have a impact on GC if we run multiples regex queries)

> If so i can create a follow up PR.

As long the code stays correct and readable, all optimizations are welcome ;).
So yes, let's discuss that over another PR.